updated css for views and updated favorites.ejs logic for db change. added new test data and updated db changes

+++ db/schema/01_schema.sql
@@ -26,7 +26,7 @@ CREATE TABLE PRODUCT(
   typeID INTEGER,
   title VARCHAR(255),
   description TEXT,
-  price INTEGER,
+  price MONEY,
   status BOOLEAN,
   img_url VARCHAR(255),
   FOREIGN KEY (userID) REFERENCES USERS(userID),
